Backend/requirements.txt
flask==3.1.0
flask-mail==0.10.0
gunicorn==23.0.0
requests==2.32.0
mysql-connector-python==9.1.0
flask-cors==3.0.10
python-dotenv
requests
gunicorn



